Integrated file read into a linked list

+++ Node.cpp
@@ -6,7 +6,7 @@ FoodItem::FoodItem(){
 
 };
 
-FoodItem::FoodItem(std::string id, std::string name, std::string description, int price){
+FoodItem::FoodItem(std::string id, std::string name, std::string description, double price){
     this -> id = id;
     this -> name = name;
     this -> description = description;
